### 2015-12-31 #24 @masalmon
Ropenaq package for R
* 1. What does this package do? (explain in 50 words or less)
It provides an interface to the OpenAQ API. OpenAQ provides open air quality data for many locations around the world, whose number is growing. The R package allows to check data availability and to download measurements from R as a dplyr data.table.
* 2. Paste the full DESCRIPTION file inside a code block (bounded by ``` on either end).

`Package: Ropenaq
Title: The Package is a R Interface to the openaq API
Version: 0.1
Author: Maëlle Salmon
Maintainer: Who to complain to <maelle.salmon@yahoo.se>
Description: See https://docs.openaq.org/
License: GPL (version 2 or later)
Suggests: knitr,
Depends: httr, lubridate, dplyr, testthat, XML`

* 3. URL for the package (the development repository, not a stylized html page)
https://github.com/masalmon/Ropenaq
* 4. What data source(s) does it work with (if applicable)?
Open AQ https://openaq.org/#/
* 5. Who is the target audience?
People that need air quality data for their studies and do not know how to access the API, or want to do the analysis in R anyway. They could be e.g. epidemiologists. I think that if the R package has users, the Open AQ platform will also get more attention and thus get even more data sources.
* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?
No, this is the first and only R package for accessing the OpenAQ API (I first asked one of the OpenAQ co-founders)
* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
* [x ] This package does not violate the Terms of Service of any service it interacts with.
* [ x] The repository has continuous integration with Travis and/or another service
* [ x] The package contains a vignette
* [ x] The package contains a reasonably complete readme with devtools install instructions
* [x ] The package contains unit tests
* [ x] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
Yes, as much as I can.
* [ ] Are there any package dependencies not on CRAN?
* [ x] Do you intend for this package to go on CRAN?
When it is more developped, yes, it could be usefool.
* [ x] Does the package have a CRAN accepted license?
* [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.
* 9. Please add explanations below for any exceptions to the above:
* 10. If this is a resubmission following rejection, please explain the change in cirucmstances.
