### 2016-01-11 masalmon/Ropenaq#4 @jflasher
Could be related to the way accent marks are being handled?
