### 2016-01-11 masalmon/Ropenaq#4 @masalmon
Yes I am working on encoding... I saw an error when the function measurements checks whether the location is available, and moreover it made me realize than when constructing the query I do not change the encoding. And a query for "Marszalkowska" is not going to give any result, I need to translate it to "Marsza%C5%82kowska". Much fun ahead. :-)
