### 2016-01-11 masalmon/Ropenaq#4 @jflasher
Haha. Yeah, it's a pain. I ran into a [similar issue](https://github.com/openaq/openaq-api/blob/ee2f1c56aa03ea3225521de99e2272bee38a796e/adapters/poland.js#L75-L108) a while back with the Polish data and did a manual fix until I realized there was a much easier way to do it. :(
