### 2016-01-11 masalmon/Ropenaq#4 @masalmon
I'm happy I tried to download data for Warsaw, because I would not have seen this issue when using Indian data... Well I'll try to solve the issue in the next days. I guess I can translate the encoding to ASCII for the query. Thanks for commenting!
