### 2016-01-11 masalmon/Ropenaq#4 @masalmon
I wanted to reproduce this graph https://pbs.twimg.com/media/CYOAGowW8AQs4Fy.png:large Turned out to be a good exercise!
