### 2016-01-11 masalmon/Ropenaq#4 @masalmon
> Sys.setlocale("LC_CTYPE", "french")
[1] "French_France.1252"
> URLencode("Marszałkowska")
[1] "Marszalkowska"
>
> Sys.setlocale("LC_CTYPE", "russian")
[1] "Russian_Russia.1251"
> URLencode("Marszałkowska")
[1] "Marszalkowska"
>
> Sys.setlocale("LC_CTYPE", "polish")
[1] "Polish_Poland.1250"
> URLencode("Marszałkowska")
[1] "Marsza%B3kowska"
