### 2016-01-11 masalmon/Ropenaq#4 @masalmon
I'm starting to wonder whether I can really convert the string to the right encoding myself. I'd need to know in which language the name of the location was given. But can I really expect users to input location as "Marsza%B3kowska"? This seems quite complicated.

@jflasher I guess the problem will be even more common as the platform grows (i.e. even more special characters)? If Polish is the only issue I could set the locale to Polish at the beginning of the function code and then back to the default setting at the end. :-)

Interestingly with setting the locale to Polish one can encode Vietnamese and Russian as well:
> Sys.setlocale("LC_CTYPE", "polish")
[1] "Polish_Poland.1250"
> URLencode("ngắn")
[1] "ng%E1%BA%AFn"
> URLencode("наприме́р")
[1] "%D0%BD%D0%B0%D0%BF%D1%80%D0%B8%D0%BC%D0%B5%C2%B4%D1%80"

(well I do not know if this encoding makes sense) Whereas if the locale is anything but Polish the "ł" is recognized as a "l"... See https://bugs.r-project.org/bugzilla3/show_bug.cgi?id=16377

 So maybe setting the locale to Polish is a good idea. I could add an argument language to all functions (cities, measurements, etc.) which by default would be Polish.
