### 2016-01-13 masalmon/Ropenaq#4 @masalmon
Solution I'm going to implement: adding a URL encoded version of each column for cities, countries, locations, and expect URL encoded arguments in measurements and latest. Not fun for users but they could find the URL encoded version from cities, countries and locations functions. And thus I would not need any side effect of the function such as changing the locale. Now I'm gonna implement all this.
