### 2016-01-13 masalmon/Ropenaq#4 @jflasher
Is there anything the API can do to make this easier?
