### 2016-01-13 masalmon/Ropenaq#4 @masalmon
No thanks, I think it will work this way. Well only experience will tell us whether it really bother users but I think only URL encoded strings are universal and as long as I can translate them from cities() and locations() anyone can find the right URL encoded name to give to a city.

I just wonder how the Python package deals with it, but it may be a R-only problem.
