### 2016-01-13 masalmon/Ropenaq#4 @jflasher
@dolugen @dhhagan have you run into anything along these lines in your projects?
