### 2016-01-13 masalmon/Ropenaq#4 @masalmon
I think I have now solved the issue for the R package. I updated the README.
