### 2016-01-14 masalmon/Ropenaq#4 @dhhagan
It is an 'issue' I guess with the python wrapper. Really it's just an encoding issue that the user can fix by encoding it the same way you have it stored in the database. I assume it's utf-8?
