### 2016-01-15 masalmon/Ropenaq#4 @dhhagan
@jflasher The way I've always 'gotten around' any issues like this is to provide a `self_url` when returning the information..also helps quite a bit with cacheing. For example..when you make the `cities` api call..it would return all of the cities, with their urls as one of the meta fields..ie

    {
      'field1': answer1,
      'self-url': https://api.openaq.org/api/cities/[utf-8 encoded name here]/,
      ...
    }

That way, you can always just grab that from the response and use it. Is that something that you would be okay with adding? I can try to poke around later and send a PR. I'm in India now though..so internet is patchy at best.
