### 2016-01-15 masalmon/Ropenaq#4 @masalmon
Regarding the R package, I now also expect the user to provide URL-encoded arguments. To help with this, for the functions that output cities/locations, I've added a column giving their URL-encoded versions. I could do a query for Warsaw without problem, which was a good test. However, I won't close this issue as long as the conversation is going on. ;-)
