### 2016-01-22 masalmon/Ropenaq#5 @masalmon
Vignette with maps not building
due to a current problem in the API.
