### 2016-01-25 masalmon/Ropenaq#5 @masalmon
See
https://github.com/openaq/openaq-api/issues/165
and https://github.com/openaq/openaq-api/issues/166
Everything else seems to work fine.
