### 2016-01-29 #24 @ateucher
@masalmon it was my pleasure - I really enjoyed doing it :) If there's anything you run into that you need help with, ping me in an issue and/or on twitter. Also, I am definitely not an expert on `httr` and parsing json content, so if any other rOpenSci people have advice that is better than mine please chime in (@sckott?)
