### 2016-01-30 #24 @sckott
nice work on the review @ateucher

@masalmon  happy to answer any `httr` questions - one that comes to mind right now:

httr `v1.1.0` has some changes https://cran.rstudio.com/web/packages/httr/

- `content()` warns now if you don't set encoding, so use the `encoding` parameter when calling `content()`
