### 2016-01-31 #24 @aammd
# review of Ropenaq

I really like this package! Its an elegant means of accessing a very interesting (and large!) online database. I have a few pieces of advice that I think will make the code more modular and easy to develop as it grows. And grow it will, because it is quite useful indeed!

## general

i had quite a challenge to try to run this package! I am using Linux (Ubuntu 14.04), and for the longest time I could not get R to talk to the API. This was extra frustrating, because i _was_ able to use the API from `curl` at the command line! I finally found the answer by posting [this StackOverflow Question](http://stackoverflow.com/questions/35055715/ssl-connect-error-in-httr-curl). h/t @richfitz for his assistance.

To correct the problem I had to run

```r
apt-get install libcurl4-openssl-dev
```

I also uninstalled then reinstalled `curl`:

```
install.packages("curl")
```

You can apparently find out if your user will be having this problem by running `curl::curl_version()$ssl_version`, and if it answers `GnuTLS`, then perhaps there should be a warning or a message directing them to this advice.

## Function reviews

### general thoughts on API functions

I really like how you've got five functions, and each maps to the five different API queries that are available. I was thinking that the names of functions might be a bit similar to functions in other packages that also use geography. What do you think of prefixing every function with a package-specific string? Perhaps something like `aq_` before all the user-facing functions (i.e. `countries()` becomes `aq_countries()`). This is similar to another rOpenSci package, `geonames`, which uses `GN` (as in `GNcities()`). This has the added benefit of playing very nicely with Rstudio, which will show all the package functions as completions when users type `aq_` and hit tab.

I'd suggest using `httr` convenience functions more. This will help when you are assembling queries and checking responses:
    - assembling queries would be easier if you use the `query` argument accepted by `httr::GET()`. For example, in `cities.R`, line `24`, you could use this approach. It will build the API call for you, and handle the encoding too:

```r
URL <- "https://api.openaq.org/v1/cities"

indonesia <- httr::GET(url = URL, query = list(country = "ID"))
```

Status codes are also easily parsed with `httr::status_code()`. For example, instead of `grepl("Gateway time-out", toString(contentPageText))` , you could use `httr::status_code(page)` to find out if 200 (everything is OK), 504 ("Gateway time-out") etc. Then, you could write some logic to stop the functions with warnings, or messages, etc as appropriate.

For parsing responses, I'd suggest you try out a two step approach - first use `grep` to identify if the content of the response is json. Then use `jsonlite::fromJSON()` to parse it. That is probably simpler and more robust than parsing by hand (e.g. as you do in `cities.R` `L30` to `L34`:

```r
## confirm that the response is what you expect:
grepl("json", headers(page)$`content-type`) # is there a better way?

## parsing content
contentPageText <- httr::content(page, as = "text")
jsonlite::fromJSON(contentPageText)

```

### `countries()`
This function looks great to me, besides the general tips above I have nothing to add. The help file might mention that the codes for countries follow the ["ISO 3166-1 alpha-2" specification](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2). (er, that is, assuming that they actually do). In a future version of the package, perhaps you could include the package [`countrycode`](https://github.com/vincentarelbundock/countrycode) to allow more flexible naming of countries.

###  `cities()`

I really like how you are checking for available countries! I was noticing that you check to see if a country is available in _all_ datasets, and I suggest you write a little country-checking function for the purpose. Perhaps something like:

```r
check_country <- function(x){

  ctrys <- as.character(countries()$code)

  is_country <- x %in% ctrys

  if(is_country){
    x
  } else {
    msg <- paste0("x must be one of ", paste0(ctrys, collapse = ", "))
    stop(msg)
  }
}

check_country("CA")
```

or perhaps something via `assertthat` :

```r
assertthat::not_empty(country)
ctrys <- as.character(countries()$code)
assertthat::validate_that(country %in% ctrys)
```

If you rely on `httr` to handle proper URL encoding, perhaps you could do away with the `cityURL` column in this output? I think it could be distracting for users to see the country information appearing twice.

#### `Ropenaq::locations`

If `Ropenaq` is not loaded, the example for this function doesn't work:

```r
Ropenaq::locations(city='Houston', parameter='co')
Error in Ropenaq::locations(city = "Houston", parameter = "co") :
  could not find function "%>%"
```

you could fix this by adding:
```
importFrom dplyr `%>%`
```
to the function.

You have a large chunk of if-else statements devoted to confirming that the inputs are actually available. You could accomplish a very similar result using `match.arg`:

`match.arg("pm25", c("pm25", "pm10", "so2","no2", "o3", "co", "bc")`

In fact, a lot of this function (Lines 34 to 154) is about checking the inputs. You could make it more succinct by doing something like this:

1 make a list that grabs all the optional arguments (`arglist <- list(parameter = parameter, has_geo = has_geo, value_from = value_from, value_to = value_to` etc)
2 get rid of any that are null , e.g. by doing `Filter(arglist, f = Negate(is.null))`
3 check the remaining ones (using `match.arg` or `assertthat`)
4 if they are OK, pass them `httr::GET`

**encoding**
Wow, you're getting _lots_ of different encodings in your results from the API! I ran the following:

```
brz <- Ropenaq::locations(country = "BR")
brz$location %>% stringi::stri_enc_mark(.)
```

And I found that you've got a *mix* of UTF-8 and ASCII encoding! I'm afraid I'm no expert in this stuff. I was looking at [this SO question](http://stackoverflow.com/questions/23699271/force-character-vector-encoding-from-unknown-to-utf-8-in-r) to try and figure out how it works. There should be a way to change the whole vector to the same encoding. Perhaps you could write the file out to a `tempfile()`, then read it back in?

Also, I know that this is probably the fault of the API and not your package, but the places named in `brz$location` are *not* places in city of Sao Paulo, but cities in the _state_ of Sao Paulo. Yet the resulting data frame has "Sao Paulo" listed for all rows, in the column "city".

### Vignettes

#### `maps.Rmd`
I wasn't able to actually get the vignette to build, until I went in and removed "NL".

When running the `for` loop at `L31` to `L33`, you're using `c()` to grow a vector. That can have the unfortunate effect of slowing down your computation. I think it could be faster (and simpler) to grab all the location info like this:

```r
country_codes <- Ropenaq::countries()$code

## create a list to hold output
locations_country <- vector(mode = "list", length = length(country_codes))
names(locations_country) <- as.character(country_codes)

for (country in country_codes) {
  print(country)
  locations_country[[country]] <- Ropenaq::locations(country = country)
}

dplyr::bind_rows(locations_country)
```

Rather than plucking out the lat and long columns, this downloads the location info and simply combines all the outputs with `dplyr::bind_rows()`.

I only get this "Gateway time-out, but try again in a few minutes." error when I look for `country=NL`. Perhaps you could add `tryCatch` to this for loop, so it doesn't automatically fail when only one country does not work? `tryCatch()` could also go inside `Ropenaq::locations()` itself, to prevent it from failing in situations like these.

That said, I wonder if it is worth actually downloading this data every time the vignette is built. Could you instead download a version of the data and place it in `data/`, or perhaps `R/sysdata.rda`? The downloading code could still serve as a useful example (though not run, you could set `eval=FALSE`), and you'd still get your (very nice!) map

ggplot2 code: no need to add theme settings twice.

The resulting map is _exquisite_ and really highlights just how much data is available from this API! I can see why you wanted to make it :smile:

#### `graphics.Rmd`

This vignette is quality: succinct, useful and attractive! Perhaps a more descriptive name for this (and other) vignettes? I think users are likely to have other packages with vignettes called `graphics` (or come to think of it, `maps`).

####  `ropenaq-vignette.Rmd`

Great intro. I'd like it to be a little longer, including some general background on what is air quality, how is it measured, who are these people over at the openaq project. A link to their website (in addition to the api specification) would be helpful.

## DESCRIPTION

A few small changes to this file would be helpful (and avoid those nasty notes on `R CMD CHECK`):

* replace "Maintainer" and "Author" with `Authors@R: person("Maelle", "Salmon", email = "maelle.salmon@yahoo.se", role = c("aut", "cre"))`
* `GPL (version 2 or later)` -> `GPL(>=2)`
* I got the note "Description should contain at least one complete sentence."
* I couldn't find where you use the `XML` package. So you could remove it from here?
* `Imports` is usually where you put the packages you need, rather than `Depends`

## MISC

* `develop/` folder. This looks like some code that you're thinking of including either in the package or in some vignettes in the future (for example, I couldn't find `findMissing2015()` used anywhere outside of `develop/finding_proportion_missing.R`). I'd suggest working on these things on a development branch, just to keep the package a bit simpler.
* i'd remove all the references to `lintr`, since that just tidies your code and doesn't affect the package at all.
