### 2016-02-03 #24 @masalmon
Ropenaq -- point-to-point answer to the reviewers
=================================================

Thanks again for your very helpful and thorough reviews, @ateucher and @aammd ! When I'm a (R) grown up I want to be a Ropensci package reviewer too and make people happy and improve their code :smiley:

I'll answer to both of you in the same answer to avoid repeating myself. I was not sure how to structure my answer so I hope it'll be easy to read -- I thought that having a remark in italic/answer would be too cumbersome. Both reviews were well structured and easy to read! Please tell me if I missed something.

General comments
================

* I was surprised too that there was no ropenaq-R-package yet but the Open AQ platform is fairly new. I asked them if they had any plan to add an R package before starting to write one because writing an existing package = :grimacing:

* Regarding the API, @jflasher and @RocketD0g are very patient and helpful so it's been a pleasure to work with them, although the API is not mature.

* If you know air quality scientists who have suggestions for new data sources for the API, you can give them this link: <https://github.com/openaq/openaq-fetch/issues> where they can open an issue for suggesting the new data source.

* I think I have fixed my description file and package structure for passing R check (except a warning for the vignettes that do not have the corresponding html -> is this an issue?). I only had two files in develop/, and now have a develop branch, I will see if I use it: one of the two files has already become a vignette for presenting how to use the R package openair with Open AQ data.

* I have added @aamd installation problem shooting in the README. It must have been quite frustrating to have to deal with it! Thanks for spotting this! Is the README the best place for such advice?

rOpenSci Packaging Guidelines
=============================

* I'm keeping the change of name (to ropenaq) for the end. Is there any clever way to change the name of the package in all files? Could it be problematic for the github repo (although it is only a change of case)?

* The names of the functions now all start with "aq_".

* I now use Imports instead of Depends.

DESCRIPTION file
================

I have made all changes.

Dependencies
============

* I have made all changes, including deleting XML. It was nice to learn about the difference between Imports and Depends. I think I've seen a very thorough article about it somewhere but always thought I was going to read it later and later never came.

* The vignette now uses ggmap. The result is not as nice as with OpenStreetMaps but I agree on its being an heavy dependency, and I mention the possibility to use OpenStreetMaps.

* I know import "%>%" from dplyr so that `Ropenaq::aq_locations(city='Houston', parameter='co')` even if the package is not loaded.

Function Usage and Documentation
================================

* I have changed all return descriptions so that the data.frames are presented as such.

* In all help files there is now a list of the columns with their meaning and class.

* In the help file of country I have written the ISO specification of the country code and that only the code, not the name, can be used. Is this clear enough or should I write this in all help files? In all help files the argument country is described as "Limit results by a certain country – a two-letters codem see countries() for finding code based on name.".

* I have tried to better explain country/city/location in the details section of the `aq_locations` function. But I'm not sure it's clear.

* I think that the possibility to query several parameters, or even several countries, should be a feature in a future version. It might be very useful indeed! I also need to find a good way to deal with very big `aq_measurements` query. I've added the `page` argument that I had somehow neglected, and a warning for when the number of found measurements is higher than limit. The warning returns the number of pages to be queried (this warning is printed no matter the value of `page`). I am not sure I should do the loop inside the function because if the user e.g. writes `aq_measurements()` it would last forever, so I prefer let the user write their own loops for now. Any suggestion? Is this something for a future version or should it be dealt with now? Should I add a verbose argument for turning warnings off?

* I am now more precise in the description of the `date_from` and `date_to` arguments.

* The latest endpoint of the API does not have any limit argument (I had to ask).

* Regarding encoding / having the URL-encoded columns for city and location, as I said in a previous comment it was the easiest solution I found for not getting any locale problem and helping the user make queries for e.g. Poland, but I'll be happy to hear any suggestions. I've noticed that if I paste a city name with polish accents in my R session, whether at home (French locale) or at work (Spanish locale), then some accents disappear which was why I created the columns.

Source Code
===========

* I check inputs very thoroughly because I've seen a remark about it in an old Ropensci package review, I think. :wink:

* I have added a utils.R file and all functions files have become much smaller. In the utils.R file there are `base_url`, `buildQuery` which returns a list of arguments for `httr::GET` after checking them, `getResults` that returns a table of results that still needs to be transformed, the three functions `functionURL`, `addCityURL` and `addLocationURL` for adding the columns with the URL-encoded names, then `addGeo`, `functionGeo` and `functionNotGeo` for adding the latitude and longitude columns, `functionTime` and `functionTime2` for transforming the columns with times, and `functionParameters` for transforming the column with the parameters names for `aq_location`. I was first happy with all functions and then I started using standard evaluation instead of non-standard evaluation... And now I have the impression that I have written very ugly code, so I'd be grateful for suggestions.

* I actually like both `importFrom` for having the "list" of functions in the NAMESPACE and `pkg::` which makes my code easier to read (at least for me). Is it ok if I let both?

* I now make a better use of `httr` and `jsonlite`:

```r
# does the query and then parses it
getResults <- function(urlAQ, argsList){
  page <- httr::GET(url = urlAQ,
                    query = argsList)

  # convert the http error to a R error
  httr::stop_for_status(page)
  contentPage <- httr::content(page, as = "text")

  # parse the data
  resTable <- jsonlite::fromJSON(contentPage)$results
  resTable <- dplyr::tbl_df(resTable)
  return(resTable)
}
```
The suggestions you both made made it all very easy! I just wonder if having a simple `stop_for_status` is user friendly enough?

* I have added a check for the size of the table resulting from the query in 4 functions (all except countries). If the table has zero rows, the function returns an empty table and a warning. This was not mentioned in the reviews but I realize I had no good message when it happened.

* @aamd, are you happy with the `buildQueries` function or do you think it would be easier to read/better if I used `assertthat`?

* I now use `tidyr` in `aq_latest`.

* I read the part about NSE/SE in Hadley Wickham's advanced R book and the vignette you pointed me to, as well as this article <http://www.numbertheory.nl/2015/09/23/using-mutate-from-dplyr-inside-a-function-getting-around-non-standard-evaluation/> but I still feel my code is really clumsy. It works (well maybe it's not even clever SE!) and it has `_` for all `dplyr` functions but it seems too complicated. I'll be happy to get any help.

Tests
=====

* There is now a test file for each context.

* I commented the failing `aq_latest` test. I promise I don't usually do this when I get a `testthat` error!

* I don't know how to test my check of the results table being empty. I don't know how to systematically write something that will yield an empty table. And now code coverage is only 94%. :laughing:

lintr
=====

* Why remove the references to lintr?

Vignettes
=========

* There are now four of them: the general one, the one with the graphics for one day in several places, the one with the map, and one for showing how to use the great `openair` package on the retrieved data.

* I have changed their filenames and titles, are they ok now?

* The vignettes now all build. Do you get any problem with them?

* The vignette creating the map now uses your code suggestion, @aamd, thanks a lot! And I saved the data in a data/ folder in the vignettes folder, is it good practice or should the data be elsewhere?

* As I mentioned earlier now this vignette uses `ggmap` instead of `OpenStreetMaps`.

* I was not sure which infos to use about air quality in the general vignette? I have added a sentence about Open AQ that I copied from their website, and a link to said website. What else would you like to see here, @aamd? Maybe I could ask @jflasher and @RocketD0g where I could find a concise description.

I hope I haven't missed anything. This review process is really very interesting. I can see that the package is getting better, and I can apply nearly everything I learnt to other R code/packages I have to write, so again, many thanks!
