### 2016-04-12 #24 @masalmon
Just giving some news about the package:

* The API recently made a few changes: now all endpoints have a limit and a page argument, so I need to change a few lines of code in the parts where I check that e.g. a country exists if the users inputs the country as argument. My package still works because I added the arguments in the functions, but it's still missing in this checking thing.

* I have started to add memoization with memoise.

* @aammd and @ateucher I'd like to add you as package authors because of all the useful feedback, is it ok?

I'll come write a comment again when I'm done with the modifications because of the API change and with memoise (a few weeks at most). I hope that the editor @sckott is ok with this "strategy"!
