### 2016-04-12 #24 @sckott
@masalmon yes, sounds good
