### 2016-04-12 #24 @noamross
Note that there is a specific semantic for adding reviewers to the Author
field of the DESCRIPTION file.  We've just added it to our recommendations
in the packaging guide:
https://github.com/ropensci/packaging_guide/blob/master/README.md#authorship

On Tue, Apr 12, 2016, 4:49 PM Scott Chamberlain <notifications@github.com>
wrote:

> @masalmon <https://github.com/masalmon> yes, sounds good
>
