### 2016-04-16 #24 @aammd
@masalmon I would be delighted to be listed as a reviewer on this package! Its a really cool package and I enjoyed studying it.

I would love to comment on the changes you've made -- I apologize for saying little in response to your [awesome & extensive revisions](https://github.com/ropensci/onboarding/issues/24#issuecomment-179176744). I confess it felt a little overwhelming! @sckott , what is the best way to go forward? Should I write another comment here?
