### 2016-04-16 #24 @sckott
> I have not changed the package name... yet? Is it compulsory?

We don't make anyone change pkg name, but we do find it easier for users if they don't have to worry about case. If you change the name, github will redirect users automatically to the new repo name `masalmon/Ropenaq` -> `masalmon/ropenaq`

> what is the best way to go forward? Should I write another comment here?

@aammd if you have time, yes please!  We realize your time is valuable, so comments don't have to be extensive
