### 2016-04-16 #24 @jennybc
Let me pre-emptively warn you about a real headscratcher re: travis:

http://stackoverflow.com/questions/31107761/encrypting-a-file-on-travis-for-a-github-repo-that-was-renamed

TL;DR manually change the travis slug in `.git/config` and save yourself hours of frustration.
