### 2016-04-17 #24 @jennybc
Based on a single repo of mine, it doesn't look like Appveyor puts info into `.git/config`, with which to confuse itself later, should you rename things. Maybe this is unique to travis.
