### 2016-04-20 #24 @masalmon
@sckott, how should I proceed?

I re-read my answers. My main doubts are now:

*  a few documentation things (whether I explain city/country/location well, and whether the few words about air quality in the readme/general vignette are enough)

* whether I should get read of all "::" if I use importFrom in the same file (I like having the "::" for seeing what I used but I know it makes computation slower... but not that much slower?)

For a future version of ropenaq my goals are:

* implement a few air quality index functions.

* making it easier to get lots of data, probably by documenting how to do this with the package versus how to go on the website and get csv dumps.

When the API is stable I want to make the package ready for CRAN submission.
