### 2016-04-21 #24 @masalmon
@sckott awesome, thank you!

The "Transfer ownership" button is in the danger zone of settings which makes me a bit nervous, ah, but it seems pretty easy.

Since I'm a member of `ropenscilabs` do I already have the right to transfer a repo there? When would the repo be transferred to `ropensci`? I'm a bit lost with all these names :grin:
