### 2016-04-21 #24 @sckott
I think you should be able to - try it and if it doesn't work I'll fix some things.

We started the 2nd organization ropenscilabs for earlier stage packages/projects - once more mature they'll move to ropensci - `more mature` is not a very specific definition - so can't give answer yet on what that means - but at some point we'll move to ropensci
